requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
praw==7.7.1
openai==1.50.0
jinja2==3.1.4
aiohttp==3.10.0
apscheduler==3.10.4
python-dotenv==1.0.1
